Reintroduce autorepr function and rename AutoRepr
I said we could reintroduce it if we needed, and we need it. (In fact,
we needed it even then, but I missed it.) We need to be able to use
autorepr as a decorator directly, as in

    @autorepr
    def __init__(self):
        ...

as well as indirectly with arguments, as in

    @autorepr(style="<>")
    def __init__(self):
        ...

We *could* do this with just the class (e.g., hijack __new__ or
__call__), but that seems unnecessarily ugly. Let the AutoRepr
descriptor just worry about reprs, and we'll bring back the function to
handle the decorator aspect.