Rewrite linkcode_resolve to be independent of current directory
There was a bug if `..` (the current directory's parent) was not a
parent directory of the easyrepr module, which it is not on Read the
Docs. There, they install easyrepr in a virtual environment but run
`make` somewhere else.

In doing this, I've removed the fallback if find_source fails. It wasn't
actually correct. It may be now that we've changed the base URL, but I'd
rather have a hard fail if we can't find the source.

Fixes #2.