Fix search for __repr__s in MRO
Our search for __repr__ methods was a bit too generous. We would find a __repr__
for B even in the following case:

    class A:
      def __repr__(self):
        ...

    class B:
      pass

In other words, getattr will search the inheritance chain for a property, even a
static property. We only want a hit for classes that define their *own* __repr__
method, because we're walking the inheritance chain manually.