Add check for compatible signature for wrapped function
The wrapped function should be callable with a single positional argument, i.e.,
`self`. We can get the signature using `inspect.signature`. Rather than parsing
the parameter list, we just attempt to bind a dummy value (`None`), which will
either succeed, in which case we ignore the `BoundArguments` object, or it will
fail with a `TypeError`, which we catch and rethrow.