Replace bottom_up parameter with top_down instance attribute
It makes sense to configure this for the mirror instance, so that
reflect_attributes can use the configured direction when iterating classes for
slots.

We've switched to top_down as the option instead of bottom_up, so that the
default can be True like hide_private. I'm still not convinced a boolean makes
sense here -- there are two options, but maybe an enum? I'm also not sure it
matters enough to spend time on this bikeshed.